+++ spec/smart_vk_api_spec.rb
@@ -1,6 +1,19 @@
 require 'spec_helper'
 
 describe SmartVkApi do
+
+  describe 'call' do
+    it { is_expected.to respond_to(:vk) }
+    it { is_expected.to respond_to(:call) }
+
+    it 'should delegate call to VK instance' do
+      vk = SmartVkApi.vk
+      allow(SmartVkApi).to receive(:vk) { vk }
+      expect(vk).to receive(:call)
+      SmartVkApi.call('users.get', {:user_ids => 'kimrgrey'})
+    end
+  end
+
   describe 'configuration' do
     it { is_expected.to respond_to(:configure) }